refactor: Refactor AWS snapshot creation and sharing
This commit refactors the code related to AWS snapshot creation and sharing. It introduces a new module called "from" that handles the creation of snapshots in the "from" AWS account. The module "to" handles the creation of snapshots in the "to" AWS account. The code changes include adding new Terraform files for the "from" module, deleting the "locals.tf" file, and modifying the "main.tf" file to use the new modules.

The changes also include the addition of the "db_snapshot.tf" and "kms_key.tf" files in the "from" module, which define the resources for creating database snapshots and managing KMS keys. The "variables.tf" file has been updated to include new variables for snapshot names.

These changes improve the organization and maintainability of the code by separating the logic for creating snapshots in different AWS accounts into separate modules.